Change organizationId to Integer
DatabaseManager expects this to be an Integer. I thought it was Long
before

+++ dspace/modules/dryad-rest-webapp/src/main/java/org/datadryad/rest/models/Organization.java
@@ -11,7 +11,7 @@ import org.codehaus.jackson.annotate.JsonIgnore;
  */
 @XmlRootElement
 public class Organization {
-    public Long organizationId;
+    public Integer organizationId;
     public String organizationCode;
     public String organizationName;
     @JsonIgnore
